Cleanup
- Moving tests for streamdeck plus to test_app_plus.py

- Fixed a bug regarding push events
